requirements.txt
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
pyyaml>=6.0.1
python-dateutil>=2.9.0.post0
